+++ website/src/components/SearchPage/DownloadDialog/DownloadDialog.tsx
@@ -5,15 +5,15 @@ import { ActiveDownloadFilters } from './ActiveDownloadFilters.tsx';
 import { DownloadButton } from './DownloadButton.tsx';
 import { DownloadForm } from './DownloadForm.tsx';
 import type { DownloadParameters } from './DownloadParameters.tsx';
-import { type DownloadOption } from './generateDownloadUrl.ts';
+import { type DownloadUrlGenerator, type DownloadOption } from './DownloadUrlGenerator.ts';
 import { routes } from '../../../routes/routes.ts';
 import type { ReferenceGenomesSequenceNames } from '../../../types/referencesGenomes.ts';
 import { formatNumberWithDefaultLocale } from '../../../utils/formatNumber.tsx';
 
 type DownloadDialogProps = {
+    downloadUrlGenerator: DownloadUrlGenerator;
     downloadParams: DownloadParameters;
     referenceGenomesSequenceNames: ReferenceGenomesSequenceNames;
-    lapisUrl: string;
 };
 
 const getDownloadButtonText = (downloadParams: DownloadParameters) => {
@@ -30,9 +30,9 @@ const getDownloadButtonText = (downloadParams: DownloadParameters) => {
 };
 
 export const DownloadDialog: FC<DownloadDialogProps> = ({
+    downloadUrlGenerator,
     downloadParams,
     referenceGenomesSequenceNames,
-    lapisUrl,
 }) => {
     const [isOpen, setIsOpen] = useState(false);
     const [downloadOption, setDownloadOption] = useState<DownloadOption | undefined>();
@@ -98,10 +98,10 @@ export const DownloadDialog: FC<DownloadDialogProps> = ({
                                     </label>
                                 </div>
                                 <DownloadButton
-                                    disabled={!agreedToDataUseTerms}
-                                    lapisUrl={lapisUrl}
+                                    downloadUrlGenerator={downloadUrlGenerator}
                                     downloadOption={downloadOption}
                                     downloadParameters={downloadParams}
+                                    disabled={!agreedToDataUseTerms}
                                     onClick={closeDialog}
                                 />
                             </div>
