+++ website/src/components/SearchPage/DownloadDialog/DownloadForm.tsx
@@ -1,7 +1,7 @@
 import { type FC, useEffect, useState } from 'react';
 
+import type { DownloadDataType, DownloadOption } from './DownloadUrlGenerator.ts';
 import { DropdownOptionBlock, RadioOptionBlock } from './OptionBlock.tsx';
-import type { DownloadDataType, DownloadOption } from './generateDownloadUrl.ts';
 import { routes } from '../../../routes/routes.ts';
 import type { ReferenceGenomesSequenceNames } from '../../../types/referencesGenomes.ts';
